Fix #2881 (Change Media Details UI)  (#2954)
* Media Details UI changed - XML file

* Onclicks added, design updated

* Removed bookmark item from media detail menu

* Update in XML
