+++ app/src/main/res/drawable/copy_to_wikitext_button.xml
@@ -9,18 +9,7 @@
         <shape
             android:shape="rectangle">
             <solid
-                android:color="@color/deleteRed"/>
-            <corners
-                android:radius="2dp" />
-        </shape>
-    </item>
-
-    <item
-        android:state_enabled="false" >
-        <shape
-            android:shape="rectangle">
-            <solid
-                android:color="@color/deleteButtonDark"/>
+                android:color="@color/button_blue"/>
             <corners
                 android:radius="2dp" />
         </shape>
