+++ app/src/main/res/layout/detail_category_item.xml
@@ -13,12 +13,9 @@
         android:background="?attr/subBackground"
         android:foreground="?attr/selectableItemBackground"
         android:gravity="center_vertical"
-        android:minHeight="@dimen/overflow_button_dimen"
-        android:padding="@dimen/quarter_standard_height"
-        android:textColor="@android:color/white"
+        android:textColor="@color/linkColor"
+        android:paddingBottom="@dimen/small_gap"
         android:textSize="@dimen/description_text_size"
-        app:drawablePadding="@dimen/small_gap"
-        app:drawableStart="@drawable/ic_info_outline_24dp"
         />
 
 </LinearLayout>
